Rename maze disposition to layout
