+++ internal/ecs/systems/score_system.go
@@ -19,7 +19,7 @@ func (ss *ScoreSystem) Update(w *ecs.World) {
 
 	for mazeEntity := range mazes {
 		mazeComp := mazes[mazeEntity].(*components.Maze)
-		maze := mazeComp.Maze
+		maze := mazeComp.Layout
 		cellSize := mazeComp.CellSize
 
 		for entity, _ := range inputControlledEntities {
